% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Shiny App for Viewing CDR Job Estimation Results}
\usage{
run_shiny_app(output_path)
}
\arguments{
\item{results}{A list of datasets to visualize.}
}
\description{
Launches a Shiny app to interactively view CDR job estimation results.
}
\examples{
run_shiny_app(results)
}
